Former-commit-id: 345550d4023340d0f319c17787c0a09793d63e27

+++ Programs/PerfomanceTests/Classes/Tests/MaterialsTest.h
@@ -42,7 +42,7 @@ private:
     static const uint32 FRAMES_PER_MATERIAL_TEST;
 
     int32 currentTestStartFrame;
-    uint64 currentTestStartTime;
+    float32 currentTestStartTime;
     uint32 currentMaterialIndex;
 
     Vector<Entity*> planes;
